+++ src/fidesops/main.py
@@ -1,15 +1,23 @@
 import logging
+from datetime import datetime, timezone
 
 import uvicorn
 from fastapi import FastAPI
+from fideslog.sdk.python.event import AnalyticsEvent
 from starlette.middleware.cors import CORSMiddleware
 
+from fidesops.analytics import (
+    in_docker_container,
+    running_on_local_host,
+    send_analytics_event,
+)
 from fidesops.api.v1.api import api_router
 from fidesops.api.v1.exception_handlers import ExceptionHandlers
 from fidesops.api.v1.urn_registry import V1_URL_PREFIX
 from fidesops.common_exceptions import FunctionalityNotConfigured
 from fidesops.core.config import config
 from fidesops.db.database import init_db
+from fidesops.schemas.analytics import EVENT
 from fidesops.tasks.scheduled.scheduler import scheduler
 from fidesops.tasks.scheduled.tasks import initiate_scheduled_request_intake
 from fidesops.util.logger import get_fides_log_record_factory
@@ -55,6 +63,15 @@ def start_webserver() -> None:
     if config.database.ENABLED:
         logger.info("Starting scheduled request intake...")
         initiate_scheduled_request_intake()
+
+    send_analytics_event(
+        AnalyticsEvent(
+            docker=in_docker_container(),
+            event=EVENT.server_start.value,
+            event_created_at=datetime.now(tz=timezone.utc),
+            local_host=running_on_local_host(),
+        )
+    )
 
     logger.info("Starting web server...")
     uvicorn.run(